backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
feedparser
python-dateutil
stripe
openai
google-auth